# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task (parent only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "200":
          $ref: "#/components/responses/Snapshot"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task (parent only)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          $ref: "#/components/responses/Snapshot"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task (parent only)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          $ref: "#/components/responses/Snapshot"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /tasks/{id}/toggle:
    post:
      operationId: toggleTask
      tags: [tasks]
      summary: Toggle a task active state (parent only)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          $ref: "#/components/responses/Snapshot"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /submissions:
    post:
      operationId: submitTask
      tags: [submissions]
      summary: Submit a task with proof (child only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionInput"
      responses:
        "200":
          $ref: "#/components/responses/Snapshot"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /submissions/{id}/review:
    post:
      operationId: reviewSubmission
      tags: [submissions]
      summary: Review a submission (parent only)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "200":
          $ref: "#/components/responses/Snapshot"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /submissions/{id}/appeal:
    post:
      operationId: appealSubmission
      tags: [submissions]
      summary: Appeal a reviewed submission (child only)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppealInput"
      responses:
        "200":
          $ref: "#/components/responses/Snapshot"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /submissions/{id}/appeal-review:
    post:
      operationId: reviewAppeal
      tags: [submissions]
      summary: Review an appeal (parent only)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppealReviewInput"
      responses:
        "200":
          $ref: "#/components/responses/Snapshot"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Snapshot:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/FamilySnapshot"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Family:
      type: object
      required:
        - id
        - name
        - parentName
        - cycleEndDate
        - cycleStartDate
      properties:
        id: { type: string }
        name: { type: string }
        parentName: { type: string }
        cycleEndDate: { type: string }
        cycleStartDate: { type: string }
    Child:
      type: object
      required: [id, familyId, name, nickname]
      properties:
        id: { type: string }
        familyId: { type: string }
        name: { type: string }
        nickname: { type: string }
    Task:
      type: object
      required:
        - id
        - familyId
        - title
        - description
        - rewardCents
        - frequency
        - assignmentType
        - assignedChildIds
        - active
        - createdAt
      properties:
        id: { type: string }
        familyId: { type: string }
        title: { type: string }
        description: { type: string }
        rewardCents: { type: integer }
        frequency: { type: string }
        assignmentType: { type: string }
        assignedChildIds:
          type: array
          items: { type: string }
        active: { type: boolean }
        createdAt: { type: string }
    Submission:
      type: object
      required:
        - id
        - taskId
        - childId
        - familyId
        - photoUri
        - status
        - rewardCentsAwarded
        - submittedForDate
        - submittedAt
      properties:
        id: { type: string }
        taskId: { type: string }
        childId: { type: string }
        familyId: { type: string }
        photoUri: { type: string }
        status: { type: string }
        rewardCentsAwarded: { type: integer }
        submittedForDate: { type: string }
        submittedAt: { type: string }
        reviewedAt: { type: ["string", "null"] }
        reviewNote: { type: ["string", "null"] }
        appealText: { type: ["string", "null"] }
        appealSubmittedAt: { type: ["string", "null"] }
    Goal:
      type: object
      required: [id, childId, title, targetCents, createdAt]
      properties:
        id: { type: string }
        childId: { type: string }
        title: { type: string }
        targetCents: { type: integer }
        createdAt: { type: string }
    StreakBet:
      type: object
      required:
        - id
        - childId
        - familyId
        - durationDays
        - startDate
        - startStreak
        - status
        - bonusPercent
        - bonusCentsAwarded
      properties:
        id: { type: string }
        childId: { type: string }
        familyId: { type: string }
        durationDays: { type: integer }
        startDate: { type: string }
        startStreak: { type: integer }
        status: { type: string }
        bonusPercent: { type: integer }
        bonusCentsAwarded: { type: integer }
        resolvedAt: { type: ["string", "null"] }
    FamilySnapshot:
      type: object
      required: [family, children, tasks, submissions, goals, streakBets]
      properties:
        family:
          $ref: "#/components/schemas/Family"
        children:
          type: array
          items: { $ref: "#/components/schemas/Child" }
        tasks:
          type: array
          items: { $ref: "#/components/schemas/Task" }
        submissions:
          type: array
          items: { $ref: "#/components/schemas/Submission" }
        goals:
          type: array
          items: { $ref: "#/components/schemas/Goal" }
        streakBets:
          type: array
          items: { $ref: "#/components/schemas/StreakBet" }
...
    TaskInput:
      type: object
      required: [title, rewardCents, frequency]
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
        rewardCents: { type: integer }
        frequency: { type: string }
        assignmentType: { type: string }
        assignedChildIds:
          type: array
          items: { type: string }
    TaskUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
        rewardCents: { type: integer }
        frequency: { type: string }
        assignmentType: { type: string }
        assignedChildIds:
          type: array
          items: { type: string }
        active: { type: boolean }
    SubmissionInput:
      type: object
      required: [taskId]
      properties:
        taskId: { type: string, minLength: 1 }
        photoUri: { type: string }
    ReviewInput:
      type: object
      required: [status, rewardCents]
      properties:
        status: { type: string }
        rewardCents: { type: integer }
        note: { type: string }
    AppealInput:
      type: object
      required: [text]
      properties:
        text: { type: string, minLength: 1 }
    AppealReviewInput:
      type: object
      required: [approved]
      properties:
        approved: { type: boolean }
        note: { type: string }
...